Create src/lib/server.ts file
This file contains common lib function used on server.
